moved and adjusted tests
